skeleton for ghost-actor proc macro
