+++ Cargo.toml
@@ -1,7 +1,9 @@
 [workspace]
 
 members = [
+  "crates/zombie_actor",
   "crates/ghost_actor",
+  "crates/ghost_actor/ghost_actor_derive",
   "crates/crypto_api",
   "crates/sodium",
   "crates/lib3h_protocol",
