+++ crates/ghost_actor/ghost_actor_derive/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "ghost_actor_derive"
+version = "0.0.10"
+authors = ["Holochain Core Dev Team <devcore@holochain.org>"]
+edition = "2018"
+description = "This is a procedural macro support crate for GhostActor: Actor System designed for ergonomic async request callbacks"
+keywords = ["holochain", "holo"]
+categories = ["asynchronous"]
+license = "Apache-2.0"
+readme = "README.md"
+documentation = "https://docs.rs/ghost_actor_derive"
+repository = "https://github.com/holochain/lib3h"
+
+[dependencies]
+ghost_actor = { version = "=0.0.10", path = ".." }
+quote = "=1.0.2"
+
+[lib]
+proc-macro = true
