+++ crates/ghost_actor/ghost_actor_derive/README.md
@@ -1,4 +1,4 @@
-# ghost_actor
+# ghost_actor_derive
 
 [![Project](https://img.shields.io/badge/project-holochain-blue.svg?style=flat-square)](http://holochain.org/)
 [![Chat](https://img.shields.io/badge/chat-chat%2eholochain%2enet-blue.svg?style=flat-square)](https://chat.holochain.net)
@@ -9,7 +9,7 @@
 
 ## Overview
 
-Actor System designed for ergonomic async request callbacks.
+This is a procedural macro support crate for GhostActor: Actor System designed for ergonomic async request callbacks.
 
 ## Contribute
