+++ crates/ghost_actor/src/code_gen/mod.rs
@@ -1,0 +1,2 @@
+mod protocol;
+pub use protocol::Protocol;
