+++ crates/lib3h/src/lib.rs
@@ -4,6 +4,7 @@ extern crate detach;
 extern crate hcid;
 extern crate lib3h_crypto_api;
 extern crate lib3h_protocol;
+extern crate lib3h_zombie_actor as lib3h_ghost_actor;
 extern crate nanoid;
 extern crate native_tls;
 extern crate tungstenite;
