+++ crates/zombie_actor/Cargo.toml
@@ -1,26 +1,24 @@
 [package]
-name = "ghost_actor"
+name = "lib3h_zombie_actor"
 version = "0.0.10"
 authors = ["Holochain Core Dev Team <devcore@holochain.org>"]
 edition = "2018"
-description = "Actor System designed for ergonomic async request callbacks"
+description = "lib3h actor request tracking module"
 keywords = ["holochain", "holo"]
 categories = ["asynchronous"]
 license = "Apache-2.0"
 readme = "README.md"
-documentation = "https://docs.rs/ghost_actor"
+documentation = "https://docs.rs/lib3h_ghost_actor"
 repository = "https://github.com/holochain/lib3h"
 
 [dependencies]
 crossbeam-channel = "=0.3.8"
-Inflector = "=0.11.4"
-lazy_static = "=1.2.0"
-log = "=0.4.8"
+detach = { version = "=0.0.10", path = "../detach" }
+lib3h_tracing = { version = "=0.0.1", path = "../tracing" }
 nanoid = "=0.2.0"
-proc-macro2 = "=1.0.2"
-quote = "=1.0.2"
 shrinkwraprs = "=0.2.1"
-syn = { version = "=1.0.5", features = [ "full", "extra-traits" ] }
+log = "=0.4.8"
 
 [dev-dependencies]
 env_logger = "=0.6.1"
+
