+++ crates/zombie_actor/README.md
@@ -1,4 +1,4 @@
-# ghost_actor
+# lib3h_ghost_actor
 
 [![Project](https://img.shields.io/badge/project-holochain-blue.svg?style=flat-square)](http://holochain.org/)
 [![Chat](https://img.shields.io/badge/chat-chat%2eholochain%2enet-blue.svg?style=flat-square)](https://chat.holochain.net)
@@ -9,7 +9,7 @@
 
 ## Overview
 
-Actor System designed for ergonomic async request callbacks.
+Lib3h actor request tracking module.
 
 ## Contribute
