+++ crates/zombie_actor/src/ghost_actor.rs
