+++ crates/zombie_actor/src/ghost_channel.rs
