+++ crates/zombie_actor/src/ghost_tracker.rs
